// src/sanity/types.ts
export type Slug = {
  _type: 'slug'
  current?: string
  source?: string
}
// ...
export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}

export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol
// Source: ./src/sanity/lib/queries.ts
// Variable: POSTS_QUERY
// Query: *[_type == "post" && defined(slug.current)][0...12]{  _id, title, slug, _createdAt, body, 	"author": author->{    name,    slug,    "image": image.asset->url  },	"mainImage": mainImage.asset->url}
export type POSTS_QUERYResult = Array<{
  _id: string
  title: string | null
  slug: Slug | null
  _createdAt: string
  body: Array<
    | {
        children?: Array<{
          marks?: Array<string>
          text?: string
          _type: 'span'
          _key: string
        }>
        style?: 'blockquote' | 'h1' | 'h2' | 'h3' | 'h4' | 'normal'
        listItem?: 'bullet'
        markDefs?: Array<{
          href?: string
          _type: 'link'
          _key: string
        }>
        level?: number
        _type: 'block'
        _key: string
      }
    | {
        asset?: {
          _ref: string
          _type: 'reference'
          _weak?: boolean
          [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
        }
        hotspot?: SanityImageHotspot
        crop?: SanityImageCrop
        alt?: string
        _type: 'image'
        _key: string
      }
  > | null
  author: {
    name: string | null
    slug: Slug | null
    image: string | null
  } | null
  mainImage: string | null
}>
// Variable: POST_QUERY
// Query: *[_type == "post" && slug.current == $slug][0]{  _id, title, slug, _createdAt, body, 	"author": author->{    name,    slug,    "image": image.asset->url  },	"mainImage": mainImage.asset->url}
export type POST_QUERYResult = {
  _id: string
  title: string | null
  slug: Slug | null
  _createdAt: string
  body: Array<
    | {
        children?: Array<{
          marks?: Array<string>
          text?: string
          _type: 'span'
          _key: string
        }>
        style?: 'blockquote' | 'h1' | 'h2' | 'h3' | 'h4' | 'normal'
        listItem?: 'bullet'
        markDefs?: Array<{
          href?: string
          _type: 'link'
          _key: string
        }>
        level?: number
        _type: 'block'
        _key: string
      }
    | {
        asset?: {
          _ref: string
          _type: 'reference'
          _weak?: boolean
          [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
        }
        hotspot?: SanityImageHotspot
        crop?: SanityImageCrop
        alt?: string
        _type: 'image'
        _key: string
      }
  > | null
  author: {
    name: string | null
    slug: Slug | null
    image: string | null
  } | null
  mainImage: string | null
} | null